requirements.txt
web3db @ git+https://github.com/timertimertimer/web3db.git
aiosqlite
web3
eth_account
python-dotenv
better_automation
better_proxy
aiohttp
aiohttp_proxy
aptos_sdk
loguru
apscheduler
aiofiles